requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage
pymysql
functions-framework
cloud-sql-python-connector